Fix the windows build

+++ src/types.h
@@ -28,9 +28,9 @@
 #define FLAG_SET(x,f)			((x) |= (f))
 #define FLAG_CLEAR(x,f)			((x) &= ~(f))
 #define FLAG_IS_SET(x,f)		((bool) (((x) & (f)) == (f)))
-#define FLAG_SET_CLEAR(x,f,c)		((x) = ((x) & ~(f)) | (-(c) & (f)))
-#define FLAG_SET_CLEAR_U8(x,f,c)	((x) = (uint8_t) (((x) & ~(f)) | (-(c) & (f))))
-#define FLAG_SET_CLEAR_U16(x,f,c)	((x) = (uint16_t) (((x) & ~(f)) | (-(c) & (f))))
+#define FLAG_SET_CLEAR(x,f,c)		((x) = ((x) & ~(f)) | (-(int)(c) & (f)))
+#define FLAG_SET_CLEAR_U8(x,f,c)	((x) = (uint8_t) (((x) & ~(f)) | (-(int)(c) & (f))))
+#define FLAG_SET_CLEAR_U16(x,f,c)	((x) = (uint16_t) (((x) & ~(f)) | (-(int)(c) & (f))))
 
 #define MASK_IS_SET(x,m,v)		((bool) (((x) & (m)) == (v)))
